Self-project-main/backend/requirements_supabase.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
pydantic==2.12.5
email-validator==2.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
supabase==2.10.0
httpx>=0.25.0
